refactor(arch-install):  rewrite all code and some new options
Now it is _actually_ written with bash (not weird, `awk` all over the place) 

* refactor: `iwctl` section.

* fix: remove `^[[0m` from the start of the iwctl output.

See https://github.com/UserIsntAvailable/scripts/pull/2#discussion_r1088563675
for more info.

* refactor: `InlineMap` iwctl output lines instead of creating new array.

* feat: add `timedatectl` and `fdisk` sections.

Even if a disk is selected on the fdisk section, no contents are gonna be
removed.

* chore: rephrase some TODO comments.

* feat: add `parted` section.

Also include `mkfs` and `mount` ones.

* fix(parted): disk aligment issues.

* docs: add comment for partition aligment information.

* feat: implement all the missing functionality.

Done with this. There are a couple of `FIX`me to clean up, but after that
everything should just work.

* refactor: use `lsblk` instead of `fdisk`

* refactor: move `sgdisk` instead of `parted`

This simplifies the partition creation, since I don't need to care about
partition alignment anymore (it is done automatically by `sgdisk`).

* fix: partition's path are correctly set

* fix(lsblk): print `path` instead of `name`

* fix: mount `/mnt/home` if the user splits root and home

* feat(lsblk): ignore `rom|loop|airoot` devices

* feat(sgdisk): use `--clear` to delete all disk data

* chore: some minor stylistic changes

* fix: bash syntax parser is not that good...

Treessitter parsed the code correctly with the `# ts-inject: bash` after the
heredoc name, but when running the script it was not running the redirect?